Pre-populate cuzebox directory on external storage with Bootloader.hex
This will create a cuzebox directory on the external storage if it
doesn't exist, and extract each file in the apk's assets directory into
it, only if that file doesn't already exist there.

Includes a blank EEPROM, except with the Boot to MENU option
pre-selected to avoid confusion after flashing the first game.